+++ easybatch-core/src/main/java/io/github/benas/easybatch/core/impl/NoOpRecordReader.java
@@ -58,7 +58,7 @@ class NoOpRecordReader implements RecordReader {
     /**
      * {@inheritDoc}
      */
-    public Long getTotalRecords() {
+    public Integer getTotalRecords() {
         return null;
     }
