+++ easybatch-tutorials/src/main/java/io/github/benas/easybatch/tutorials/jms/GreetingJmsReader.java
@@ -67,7 +67,7 @@ public class GreetingJmsReader implements RecordReader {
     }
 
     @Override
-    public Long getTotalRecords() {
+    public Integer getTotalRecords() {
         //undefined, cannot be calculated in advance
         return null;
     }
